src: lib: mavlink: [clippy] Remove uneeded cast

+++ src/lib/stream/sink/types/zenoh_message.rs
@@ -19,7 +19,7 @@ impl Timestamp {
         let now = chrono::Utc::now();
         Self {
             sec: now.timestamp() as u32,
-            nsec: now.timestamp_subsec_nanos() as u32,
+            nsec: now.timestamp_subsec_nanos(),
         }
     }
 }
